+++ .tests/tests/framework-wordpress.sh
@@ -13,7 +13,7 @@ DVLBOX_PATH="$( cd "${SCRIPT_PATH}/../.." && pwd -P )"
 . "${SCRIPT_PATH}/../scripts/.lib.sh"
 
 RETRIES=20
-DISABLED_VERSIONS=("8.2")
+DISABLED_VERSIONS=("")
 
 
 echo
@@ -56,18 +56,23 @@ VHOST="my-wordpress"
 # Create vhost dir
 create_vhost_dir "${VHOST}"
 
-
+# Switch to an earlier Wordpress version for older PHP versions
+if [ "${PHP_VERSION}" = "5.6" ]; then
+	WP_BRANCH=6.2.4
+else
+	WP_BRANCH="$(git ls-remote --tags https://github.com/WordPress/WordPress | sed 's/^.*tags\///g' | grep -E '^[.0-9]+$' | tr '-' '~' | sort -V | tail -1)"
+fi
 # Download Wordpress
 run "docker-compose exec --user devilbox -T php bash -c ' \
-	git clone https://github.com/WordPress/WordPress /shared/httpd/${VHOST}/wordpress \
+	git clone --depth=1 --single-branch --branch=${WP_BRANCH} https://github.com/WordPress/WordPress /shared/httpd/${VHOST}/wordpress \
 	&& ln -sf wordpress /shared/httpd/${VHOST}/htdocs'" \
 	"${RETRIES}" "${DVLBOX_PATH}"
 
 # Switch to an earlier Wordpress version for older PHP versions
-if [ "${PHP_VERSION}" = "5.3" ] || [ "${PHP_VERSION}" = "5.4" ] || [ "${PHP_VERSION}" = "5.5" ]; then
+if [ "${PHP_VERSION}" = "5.6" ]; then
 	run "docker-compose exec --user devilbox -T php bash -c ' \
 		cd /shared/httpd/${VHOST}/wordpress \
-		&& git checkout 5.1.3'" \
+		&& git checkout 6.2.4'" \
 	"${RETRIES}" "${DVLBOX_PATH}"
 # Checkout latest git tag
 else
@@ -75,7 +80,6 @@ else
 		cd /shared/httpd/${VHOST}/wordpress \
 		&& git checkout \"\$(git tag | sort -V | tail -1)\"'" \
 	"${RETRIES}" "${DVLBOX_PATH}"
-
 fi
 
 # Setup Database